pathways node metadata initial implementation - add identifiers url, identifiers url generation function - delete bloated code - pathways node metadata model no longer needed - overall code is simplified - TODO write tests
