+++ src/server/routes/pathways/generate-pathway-json/biopax-metadata/index.js
@@ -1,237 +1,145 @@
 const fs = require('fs');
-const metadataParser = require('./metadata-parser');
-const { populateMetadata } = require('./simple');
 const _ = require('lodash');
 
-var biopaxMap = null;
+let extractBiopaxMetadata = biopaxJsonEntry => {
+  let databaseIds = _.get(biopaxJsonEntry, 'dbIds', []);
+  let entRefEl = _.get(biopaxJsonEntry, 'entRefEl', []);
 
-/**
- *
- * @param {*} biopaxElement Metadata for an entity from BioPAX
- * @param {*} nodeType Is this a xref or entity?
- */
-function collectEntityMetadata(biopaxElement, nodeType = 'default'){
+  let type = _.get(biopaxJsonEntry, '@type', '');
+  let datasource = _.get(biopaxJsonEntry, 'dataSource', '');
+  let comments = _.get(biopaxJsonEntry, 'comment', []);
 
-  let result = [];
-  //Collect relevant data on this node, and push it into output
-  if (nodeType !== 'Reference') {
-    //Get type
-    let type = biopaxElement['@type'];
-    if (type) { result.push(['Type', type]); }
 
-    //Get data source
-    let dataSource = biopaxElement['dataSource'];
-    if (dataSource) { result.push(['Data Source', dataSource]); }
+  let entryDisplayName = _.get(biopaxJsonEntry, 'displayName', '');
+  let entryNames = [].concat(_.get(biopaxJsonEntry, 'name', []));
+  let entryStdName = _.get(biopaxJsonEntry, 'standardName', '');
 
-    //Get display name
-    let dName = biopaxElement['displayName'];
-    if (dName) { result.push(['Display Name', dName]); }
+  let entRefNames = _.get(entRefEl, 'name', []);
+  let entRefStdName = _.get(entRefEl, 'standardName', '');
+  let entRefDisplayName = _.get(entRefEl, 'displayName', '');
 
-    //Get Comments
-    let comment = biopaxElement['comment'];
-    if (comment) { result.push(['Comment', comment]); }
 
-    //Get Names
-    let name = biopaxElement['name'];
-    if (name) { result.push(['Names', name]); }
+  let synonyms = ( 
+    _.uniq( 
+      _.flatten( [entryNames, entRefNames] ) 
+    ).filter( name => !_.isEmpty( name ) )
+  );
 
-    //Get Standard Name
-    let sName = biopaxElement['standardName'];
-    if (sName) { result.push(['Standard Name', sName]); }
+  let standardName = entryStdName || entRefStdName || '';
+  let displayName = entryDisplayName || entRefDisplayName || '';
 
-    //Get Chemical Formula
-    let formula = biopaxElement['chemicalFormula'];
-    if (formula) { result.push(['Chemical Formula', formula]); }
-  }
+  return {
+    comments,
+    synonyms,
+    datasource,
+    type: type.includes(':') ? type.split(':')[1] : type,
+    standardName,
+    displayName,
+    databaseIds
+  };
+};
 
-  //Get database id
-  let db = biopaxElement['db'];
-  let id = biopaxElement['id'];
-  if (db && id) { result.push(['Database ID', [db, id]]); }
+// transform biopaxJsonText into a consolidated js object
+let biopaxText2ElementMap = biopaxJsonText => {
+  let rawMap = new Map();
+  let elementMap = new Map();
+  let xRefMap = new Map();
 
-  return result;
-}
+  let biopaxElementGraph = JSON.parse(biopaxJsonText)['@graph'];
+  let externalReferences = [];  
 
-/**
- *
- * @param {*} entity Metadata for an entity in the network
- * @returns Tree array containing all metadata about the given entity
- */
-function buildBiopaxTree(entity) {
-  // console.log(entity);
-  // console.log('\n\n\n\n');
-  let result = [];
-
-  //make sure the metadata exists
-  if (!(entity))
-    return;
-
-  //Get the node id, as defined in getProcessedBioPax
-  let id = entity['pathid'];
-  if (!(id)) { id = entity['about']; }
-  if (!(id)) { return; }
-
-  //get BioPAX metadata for this node
-  let collectedData = collectEntityMetadata(entity);
-
-  //Collect Data for cross-references
-  let xrefList = [];
-  let xref = entity['xref'];
-  let eref = entity['entityReference'];
-  if(xref)
-    if(typeof xref === 'string')
-      xrefList = _.concat(xrefList,[xref]);
-    else
-      xrefList = _.concat(xrefList,xref);
-  if(eref)
-    xrefList = _.concat(xrefList,[eref]);
-
-  processXrefs(xrefList,eref,collectedData);
-
-  result.push([id,collectedData]);
-  return result;
-}
-
-function processXrefs(xrefList,eref,collectedData){
-  let erefXrefList = [];
-
-  for (let i = 0; i < xrefList.length; i++) {
-    //Check if this is a cross-reference or entity reference
-    //(- the last item (if eref is defined) is treated in a special way that
-    //TODO: saves a couple of lines but sacrifices clarity...)
-    let keyName = 'Reference';
-    if (i == (xrefList.length - 1) && eref)
-      keyName = 'EntityReference';
-
-    //Get Referenced element and make sure it's valid
-    let refElement = getByNodeId(xrefList[i], biopaxMap);
-
-    if (!(refElement)) {
-      continue;
+  biopaxElementGraph.forEach( element => {
+    if( _.has(element, '@id') ){
+      rawMap.set(element['@id'], element);
     }
 
-    //create list of xrefs for eref
-    if(keyName === "EntityReference") {
-      let erefXref = refElement['xref'];
-      if(erefXref)
-        if(typeof erefXref === 'string')
-          erefXrefList = _.concat(erefXrefList,[erefXref]);
-        else
-          erefXrefList = _.concat(erefXrefList,erefXref);
+    if( _.has(element, 'xref' ) ){
+      externalReferences.push(element);
+    }
+  });
+
+  // extract db ids from specific elements that have 'db' and 'id' fields
+  externalReferences.forEach( element => {
+    let xrefs = [].concat(_.get(element, 'xref', []));
+
+    xrefs.forEach( xrefId => {
+      if( rawMap.has( xrefId ) ){
+        let { db, id } = rawMap.get( xrefId );
+
+        if( db != null && id != null ){
+          xRefMap.set(xrefId, { k: db, v: id });
+        }
+      }
+    });
+  });
+
+  biopaxElementGraph.forEach( element => {
+    let entityReference = _.get(element, 'entityReference', null);
+    let xrefIds = [].concat( _.get(element, 'xref', []) );
+    let elementId = _.get(element, '@id');
+    let dbIds = {};
+    let entRefEl = rawMap.get( entityReference );
+
+    if( entRefEl != null ){
+      let entRefXrefs = _.get(entRefEl, 'xref', []);
+      xrefIds = xrefIds.concat( entRefXrefs );
     }
 
-    //Collect data and add to tree array
-    collectedData.push([keyName,collectEntityMetadata(refElement,keyName)]);
-  }
+    // consolidate all the db ids from all the xrefs relevant to this element
+    xrefIds.filter( xrefId => xrefId != null ).forEach( xrefId => {
+      let { k, v } = xRefMap.get( xrefId );
 
-  //Also collect data for entity reference's cross-references
-  //almost identically to above
-  for(let i=0;i<erefXrefList.length;i++){
-    let keyName = 'Reference';
-    let refElement = getByNodeId(erefXrefList[i], biopaxMap);
-    if (!(refElement)) {
-      continue;
+      if( dbIds[k] != null ){
+        dbIds[k] = dbIds[k].concat(v);
+      } else {
+        dbIds[k] = [v];
+      }
+    });
+
+    // each 'element' does not contain all of the data we need, it 
+    // is scattered across various xref elements and entityReference elements.
+    // we merge all this data into one object for easy processing
+    elementMap.set( elementId, _.assign( element, { dbIds, entRefEl } ) );
+  });
+
+  return elementMap;
+};
+
+
+let getBiopaxMetadata = ( cyJsonNodes, biopaxJsonText ) => {
+  let bm = biopaxText2ElementMap( biopaxJsonText );
+  let cyJsonNodeMetadataMap = {};
+
+  cyJsonNodes.forEach( node => {
+    let nodeId = node.data.id;
+    let altPCId = nodeId.substring(0, nodeId.lastIndexOf('_'));
+
+    // weird legacy hack to get extra metadata for certain nodes that have PC prefixes
+    if( bm.has( nodeId ) ){
+      cyJsonNodeMetadataMap[nodeId] = extractBiopaxMetadata( bm.get(nodeId) );
+    } else {
+      if( bm.has( altPCId ) ){
+        cyJsonNodeMetadataMap[nodeId] = extractBiopaxMetadata( bm.get(altPCId) );
+      }
     }
+  });
 
-    collectedData.push([keyName,collectEntityMetadata(refElement,keyName)]);
-  }
-  // console.log(erefXrefList);
-  // console.log(xrefList);
-}
+  return cyJsonNodeMetadataMap;
+};
 
-/**
- * Gets an object from the given map by biopax URI or Cy node id;
- * node id is mapped onto the biopax URI using several rules
- * (knowing how the SBGN xml ids and CyJson node ids were generated).
- *
- * @param id String representing a BioPAX absolute URI or Cy node id
- * @param  map Map/Object - biopax URI (string) to some objects map (e.g., to elements or metadata)
- * @returns matching value from the Map (if it exists) or null
- */
-function getByNodeId(id, map) {
-  // shortcut
-  if(!id || !map)
-    return null;
 
-  let bpe = (map instanceof Map) ? map.get(id) : map[id];
-  if (!bpe) {
-    //remove the last underscore and everything after, and search again (as generics/complex component ids
-    //are sometimes like ..Protein_4c4358ebaabf0f39e1e5325a4178d931_830703244edf74707f84842e080f965d)
-    id = id.substring(0, id.lastIndexOf("_"));
-    bpe = (map instanceof Map) ? map.get(id) : map[id];
-  }
-
-  return (bpe) ? bpe : null;
-}
-
-/**
- * Builds a URI-to-biopaxElement map from a biopax JSON-LD model.
- * @param {*} biopaxJsonText String - BioPAX sub-network in JSON-LD format
- * @returns `Map` containing BioPAX JSON objects
- */
-function getBiopaxMap(biopaxJsonText) {
-
-  //parse String into JSON object, rename '@id' property to 'pathid'
-  const graph = JSON.parse(biopaxJsonText.replace(new RegExp('@id', 'g'), 'pathid'))['@graph'];
-  // console.log(JSON.parse(biopaxJsonText)['@graph']);
-  const biopaxMap = new Map();
-
-  //'pathid' is the absolute URI (URL, by design) of a biopax object, and so we use it as map key:
-  for (const element of graph) {
-    const uri = element['pathid'];
-    biopaxMap.set(uri, element);
-  }
-
-  return biopaxMap;
-}
-
-const getGenericPhysicalEntityMap = _.memoize(() => JSON.parse(
+let getGenericPhysicalEntityMap = _.memoize(() => JSON.parse(
   fs.readFileSync(__dirname + '/generic-physical-entity-map.json', 'utf-8')
 ));
 
 
-/**
- * Extracts additional BioPAX properties, for cy nodes,
- * from the corresp. JSON-LD representation
- * (these were missing from the initial SBGN-ML result, converted to CyJson.)
- *
- * @param {*} biopaxJsonText String containing BioPAX metadata for the network
- * @param {*} nodes JSON object containing all the nodes in the network
- * @returns Processed metadata for each node in the network
- */
-function getBiopaxMetadata(biopaxJsonText, nodes) {
 
-  //BioPAX metadata comes as a string, turn it into a Map
-  biopaxMap = getBiopaxMap(biopaxJsonText);
-
-  const nodeMetadataMap = {};
-
-  // console.log(JSON.stringify(JSON.parse(biopaxJsonText), null, 2));
-
-  // console.log(populateMetadata( nodes, biopaxJsonText ));
-  //Iterate through nodes in Cy network, adding metadata to each
-  nodes.forEach(node => {
-    const bpe = getByNodeId(node.data.id, biopaxMap);
-    //build the tree for metadata and add it to the map
-    let data = metadataParser(buildBiopaxTree(bpe));
-    // console.log(JSON.stringify(data, null, 2));
-    // console.log('\n\n\n\n');
-    nodeMetadataMap[node.data.id] = data;
-  });
-
-  return nodeMetadataMap;
-}
-
-/**
- *
- * @param nodes
- */
-const getGeneSymbolsForGenericNodes = nodes => {
-  const nodeGeneSynonyms = {};
-  const genericPhysicalEntityMap = getGenericPhysicalEntityMap();
+let getGenericPhyiscalEntityData = nodes => {
+  let nodeGeneSynonyms = {};
+  let genericPhysicalEntityMap = getGenericPhysicalEntityMap();
 
   nodes.forEach(node => {
-    const genericPE = getByNodeId(node.data.id, genericPhysicalEntityMap);
+    let genericPE = genericPhysicalEntityMap[node.data.id];
     let syns = _.get(genericPE, 'synonyms', []);
     if( syns == null) { syns = []; }
     nodeGeneSynonyms[node.data.id] = syns;
@@ -241,4 +149,4 @@ const getGeneSymbolsForGenericNodes = nodes => {
 };
 
 
-module.exports = {getBiopaxMetadata, getGeneSymbolsForGenericNodes};
+module.exports = { getBiopaxMetadata, getGenericPhyiscalEntityData };
